% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advserseimpact-package.R
\docType{package}
\name{advserseimpact-package}
\alias{advserseimpact}
\alias{advserseimpact-package}
\title{advserseimpact: Statistical and Practical Significance Testing for Adverse Ampact in Selection Decisions}
\description{
Convenience functions to make some common tasks with string manipulation and regular expressions a bit easier.
}
\author{
\strong{Maintainer}: Dragos Iliescu \email{dragos.iliescu@fpse.unibuc.ro}

Authors:
\itemize{
  \item George Gunnesch-Luca \email{george.gunnesch-luca@fau.de}
}

}
\keyword{internal}
