% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aicompute.R
\name{ai.compute}
\alias{ai.compute}
\title{Calculate Statistical and Practical Significance for AI Values}
\usage{
ai.compute(NFmin, NPmin, NFmaj, NPmaj)
}
\arguments{
\item{NFmin}{Value2}

\item{NPmin}{Value5}

\item{NFmaj}{Value3}

\item{NPmaj}{Value2}
}
\value{
A vector containing parameter estimates
}
\description{
Calculate Statistical and Practical Significance for AI Values
}
