% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_absdif.R
\name{ai_absdif}
\alias{ai_absdif}
\title{Calculate Absolute Difference in Selection Rates}
\usage{
ai_absdif(p1, p2)
}
\arguments{
\item{p1}{Selection rate for the first group}

\item{p2}{Selection rate for the second group}
}
\value{
The absolute difference between the selection rates of the two groups
}
\description{
This function calculates the absolute difference in selection rates between two subgroups.
}
\details{
The calculation method was referenced from:
Oswald, F.L., Dunleavy, E.M., & Shaw, A. (2017). Measuring practical significance in adverse
impact analysis. In S.B. Morris & E.M. Dunleavy (Eds.), Adverse Impact Analysis:
Understanding Data, Statistics, and Risk (p. 105). Routledge.
}
\examples{
# For men with a selection rate of 75\% and women with a selection rate of 60\%
p1 <- 0.75
p2 <- 0.60
absolute_difference <- ai_absdif(p1, p2)
print(absolute_difference)

}
