% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_chi.R
\name{ai_chi}
\alias{ai_chi}
\title{Chi-square statistic calculation}
\usage{
ai_chi(NFmin, NPmin, NFmaj, NPmaj)
}
\arguments{
\item{NFmin}{Number of non-selected individuals in the minority group}

\item{NPmin}{Number of selected individuals in the minority group}

\item{NFmaj}{Number of non-selected individuals in the majority group}

\item{NPmaj}{Number of selected individuals in the majority group}
}
\value{
The a dataframe with computed Chi-square statistic and associated probability level (Two-tailed and One-tailed).
}
\description{
This function calculates the chi-square statistic based on selected and non-selected individuals in two groups
(e.g., minority and majority).
}
\examples{
# Example data
NFmin <- 10 # Number of participants failing the selection procedure from the minority group
NPmin <- 5  # Number of participants passing the selection procedure from the minority group
NFmaj <- 30 # Number of participants failing the selection procedure from the majority group
NPmaj <- 15 # Number of participants passing the selection procedure from the majority group

# Perform the two-sample proportion test
result <- ai_chi(NFmin, NPmin, NFmaj, NPmaj)
result

}
