% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_disparity.R
\name{ai_disp}
\alias{ai_disp}
\title{Adverse Impact Disparity Calculation}
\usage{
ai_disp(NFmin, NPmin, NFmaj, NPmaj)
}
\arguments{
\item{NFmin}{Number of participants failing the selection procedure from the minority group}

\item{NPmin}{Number of participants passing the selection procedure from the minority group}

\item{NFmaj}{Number of participants failing the selection procedure from the majority group}

\item{NPmaj}{Number of participants passing the selection procedure from the majority group}
}
\value{
A dataset containing the absolute disparity (AD) and absolute impact (AI)
}
\description{
Adverse Impact Disparity Calculation
}
\examples{
# Test cases
NFmin <- 1771
NPmin <- 338
NFmaj <- 25
NPmaj <- 532

# Calculate the absolute disparity (AD) and absolute impact (AI)
disparity_result <- ai_disp(NFmin, NPmin, NFmaj, NPmaj)
print(disparity_result)

}
