% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_FET.R
\name{ai_fet}
\alias{ai_fet}
\title{Compute the Fisher's Exact Test (FET) Statistic}
\usage{
ai_fet(NFmin, NPmin, NFmaj, NPmaj)
}
\arguments{
\item{NFmin}{The number of non-favored outcomes for the minority group}

\item{NPmin}{The number of favored outcomes for the minority group}

\item{NFmaj}{The number of non-favored outcomes for the majority group}

\item{NPmaj}{The number of favored outcomes for the majority group}
}
\value{
A numeric value representing the FET statistic
}
\description{
This function calculates the Fisher's Exact Test (FET) statistic, a measure used
in adverse impact analysis. The FET is a statistical test used to determine if
there are nonrandom associations between two categorical variables.
}
\examples{
NFmin <- 25
NPmin <- 75
NFmaj <- 100
NPmaj <- 300
ai_fet(NFmin, NPmin, NFmaj, NPmaj)

}
