% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_ir.R
\name{ai_ir}
\alias{ai_ir}
\title{Computes the Impact Ratio (IR) between minority and majority groups}
\usage{
ai_ir(NFmin, NPmin, NFmaj, NPmaj)
}
\arguments{
\item{NFmin}{Number of participants failing the selection procedure from the minority group}

\item{NPmin}{Number of participants passing the selection procedure from the minority group}

\item{NFmaj}{Number of participants failing the selection procedure from the majority group}

\item{NPmaj}{Number of participants passing the selection procedure from the majority group}
}
\value{
A numeric value representing the Impact Ratio (IR) between the minority and majority groups.
}
\description{
This function calculates the Impact Ratio (IR) by dividing the selection rate of the minority group (SRmin)
by the selection rate of the majority group (SRmaj). The IR helps assess the relative difference in selection
rates between the two groups.
}
\examples{
# Example data
NFmin <- 10 # Number of participants failing the selection procedure from the minority group
NPmin <- 5  # Number of participants passing the selection procedure from the minority group
NFmaj <- 30 # Number of participants failing the selection procedure from the majority group
NPmaj <- 15 # Number of participants passing the selection procedure from the majority group

# Calculate the Impact Ratio
result <- ai_ir(NFmin, NPmin, NFmaj, NPmaj)
result
}
