% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_lanc.R
\name{ai_lancaster}
\alias{ai_lancaster}
\title{Compute the LMP Test Statistic}
\usage{
ai_lancaster(NFmin, NPmin, NFmaj, NPmaj)
}
\arguments{
\item{NFmin}{The number of non-favored outcomes for the minority group}

\item{NPmin}{The number of favored outcomes for the minority group}

\item{NFmaj}{The number of non-favored outcomes for the majority group}

\item{NPmaj}{The number of favored outcomes for the majority group}
}
\value{
A dataset representing the LMP test statistic (One-tailed only)
}
\description{
This function calculates the LMP test statistic, which is a measure used in
adverse impact analysis. The LMP test is based on the Fisher's Exact Test.
}
\examples{
#'
NFmin <- 25
NPmin <- 75
NFmaj <- 100
NPmaj <- 300
ai_lancaster(NFmin, NPmin, NFmaj, NPmaj)

}
