% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_or.R
\name{ai_odds}
\alias{ai_odds}
\title{AI Odds Ratio Calculation}
\usage{
ai_odds(NFmin, NPmin, NFmaj, NPmaj)
}
\arguments{
\item{NFmin}{Number of participants failing the selection procedure from the minority group}

\item{NPmin}{Number of participants passing the selection procedure from the minority group}

\item{NFmaj}{Number of participants failing the selection procedure from the majority group}

\item{NPmaj}{Number of participants passing the selection procedure from the majority group}
}
\value{
The odds ratio
}
\description{
This function calculates the odds ratio between two groups (a minority group and a majority group)
based on the number of participants passing or failing a selection procedure.
}
\examples{
ai_odds(30, 70, 40, 60)
}
