% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_or.R
\name{ai_or}
\alias{ai_or}
\title{Calculate Odds Ratio}
\usage{
ai_or(p1, p2)
}
\arguments{
\item{p1}{Selection rate for the first group (e.g., men)}

\item{p2}{Selection rate for the second group (e.g., women)}
}
\value{
The odds ratio, a measure of association between the selection rates of two groups.
}
\description{
This function takes the selection rates of two groups (e.g., men and women)
and calculates the odds ratio.
}
\details{
The algorithm used in this function is based on the method described in:
Oswald, F.L., Dunleavy, E.M., & Shaw, A. (2017). Measuring practical significance in adverse
impact analysis. In S.B. Morris & E.M. Dunleavy (Eds.), Adverse Impact Analysis:
Understanding Data, Statistics, and Risk (p. 105). Routledge.
}
\examples{
# For men with a selection rate of 75\% and women with a selection rate of 60\%
p1 <- 0.75
p2 <- 0.60
odds_ratio <- ai_or(p1, p2)
print(odds_ratio)

}
