% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_prm.R
\name{ai_prm}
\alias{ai_prm}
\title{Compute the sample and selection parameters and the probability to trigger the Adverse impact}
\usage{
ai_prm(NFmin, NPmin, NFmaj, NPmaj, ds = data("d"))
}
\arguments{
\item{NFmin}{Number of participants failing the selection procedure from the minority group}

\item{NPmin}{Number of participants passing the selection procedure from the minority group}

\item{NFmaj}{Number of participants failing the selection procedure from the majority group}

\item{NPmaj}{Number of participants passing the selection procedure from the majority group}

\item{ds}{Monte Carlo simulated dataset with all probability to trigger the Adverse impact}
}
\value{
A dataset with all information related to 0, 0.20, 0.50 and 0.80 effect size
}
\description{
This function calculates the overall sample size, Overall selection ratio,
Proportion of Minority Group and Minority Selection Ratio parameters and
the probability to trigger the Adverse impact related to corresponding indicator at
different effect sizes (Cohen's d 0.00, 0.20, 0.50 and 0.80).
}
\details{
The calculation method was referenced from:
Oswald, F.L., Dunleavy, E.M., & Shaw, A. (2017). Measuring practical significance in adverse
impact analysis. In S.B. Morris & E.M. Dunleavy (Eds.), Adverse Impact Analysis:
Understanding Data, Statistics, and Risk (p. 105). Routledge.
}
\examples{
# Example data
NFmin <- 10 # Number of participants failing the selection procedure from the minority group
NPmin <- 5  # Number of participants passing the selection procedure from the minority group
NFmaj <- 30 # Number of participants failing the selection procedure from the majority group
NPmaj <- 15 # Number of participants passing the selection procedure from the majority group
prm <- ai_prm(NFmin, NPmin, NFmaj, NPmaj)
print(prm)

}
