% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_adj_shortfall.R
\name{ai_sfadj}
\alias{ai_sfadj}
\title{Adjusted Shortfall}
\usage{
ai_sfadj(NFmin, NPmin, NFmaj, NPmaj)
}
\arguments{
\item{NFmin}{Number of participants failing the selection procedure from the minority group}

\item{NPmin}{Number of participants passing the selection procedure from the minority group}

\item{NFmaj}{number of participants failing the selection procedure from the majority group}

\item{NPmaj}{Number of participants passing the selection procedure from the majority group}
}
\value{
A dataset representing the adjusted shortfall (SFadj) calculated as the
shortfall (SF) divided by the total number of minority applicants (Nmin).
}
\description{
Adjusted Shortfall
}
\examples{
# Example data
NFmin <- 10 # Number of participants failing the selection procedure from the minority group
NPmin <- 5  # Number of participants passing the selection procedure from the minority group
NFmaj <- 30 # Number of participants failing the selection procedure from the majority group
NPmaj <- 15 # Number of participants passing the selection procedure from the majority group

# Calculate the adjusted shortfall
result <- ai_sfadj(NFmin, NPmin, NFmaj, NPmaj)
result
}
