% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_adj_shortfall.R
\name{ai_shortfall}
\alias{ai_shortfall}
\title{Adjusted Shortfall}
\usage{
ai_shortfall(NFmin, NPmin, NFmaj, NPmaj)
}
\arguments{
\item{NFmin}{Number of participants failing the selection procedure from the minority group}

\item{NPmin}{Number of participants passing the selection procedure from the minority group}

\item{NFmaj}{number of participants failing the selection procedure from the majority group}

\item{NPmaj}{Number of participants passing the selection procedure from the majority group}
}
\description{
Further documentation goes here:
}
