% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_tstat.R
\name{ai_tot}
\alias{ai_tot}
\title{Compute the marginal totals and all proportions}
\usage{
ai_tot(NFmin, NPmin, NFmaj, NPmaj)
}
\arguments{
\item{NFmin}{Number of non-selected individuals in the minority group}

\item{NPmin}{Number of selected individuals in the minority group}

\item{NFmaj}{Number of non-selected individuals in the majority group}

\item{NPmaj}{Number of selected individuals in the majority group}
}
\value{
A dataframe with Number of individuals in the minority group (Nmin),
Number of individuals in the majority group (Nmaj)
Number of individuals in the failed group (Nfai)
Number of individuals in the passed group (Npas)
Total number of individuals (Ntot)
Success ratio in minority group (SRmin)
Success ratio in majority group (SRmaj)
Total success ratio (SRtot)
Proportion of individuals with feature in minority group (Pmin),
Proportion of individuals with feature in minority group (Pmaj),
Overall proportion of individuals with the feature (Ptot)
}
\description{
This function calculates the the total sample size of the minority group and majority group,
the proportion of individuals with the feature in the minority and majority groups
}
\examples{
NFmin <- 25
NPmin <- 75
NFmaj <- 100
NPmaj <- 300
ai_tot(NFmin, NPmin, NFmaj, NPmaj)

}
