% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aicompute.R
\name{ai.compute}
\alias{ai.compute}
\title{Calculate Statistical and Practical Significance for AI Values}
\usage{
ai.compute(NFmin, NPmin, NFmaj, NPmaj)
}
\arguments{
\item{NFmin}{Number of applicants failing selection from the minority group}

\item{NPmin}{Number of applicants passing selection from the minority group}

\item{NFmaj}{Number of applicants failing selection from the majority group}

\item{NPmaj}{Number of applicants passing selection from the majority group}
}
\value{
A vector containing the results.
}
\description{
Calculate Statistical and Practical Significance for AI Values
}
