% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_chi.R
\name{ai_chi}
\alias{ai_chi}
\title{Chi-square statistic calculation}
\usage{
ai_chi(NFmin, NPmin, NFmaj, NPmaj)
}
\arguments{
\item{NFmin}{Number of non-selected individuals in the minority group}

\item{NPmin}{Number of selected individuals in the minority group}

\item{NFmaj}{Number of non-selected individuals in the majority group}

\item{NPmaj}{Number of selected individuals in the majority group}
}
\value{
The computed Chi-square statistic and associated probability level
}
\description{
This function calculates the chi-square statistic based on selected and non-selected individuals in two groups
(e.g., minority and majority).
}
\examples{
ai_chi(40, 60, 25, 75)
ai_chi(1771, 338, 25, 532)

}
