% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_ir.R
\name{ai_ir}
\alias{ai_ir}
\title{Title}
\usage{
ai_ir(NFmin, NPmin, NFmaj, NPmaj)
}
\arguments{
\item{NFmin}{Number of participants failing the selection procedure from the minority group}

\item{NPmin}{Number of participants passing the selection procedure from the minority group}

\item{NFmaj}{number of participants failing the selection procedure from the majority group}

\item{NPmaj}{Number of participants passing the selection procedure from the majority group}
}
\description{
Title
}
