% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_sf.R
\name{ai_sf}
\alias{ai_sf}
\title{Title}
\usage{
ai_sf(NFmin, NPmin, NFmaj, NPmaj)
}
\arguments{
\item{NFmin}{Number of participants failing the selection procedure from the minority group}

\item{NPmin}{Number of participants passing the selection procedure from the minority group}

\item{NFmaj}{number of participants failing the selection procedure from the majority group}

\item{NPmaj}{Number of participants passing the selection procedure from the majority group}
}
\value{
A vector containing the result
}
\description{
Title
}
