% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_ztest.R
\name{ai_ztest}
\alias{ai_ztest}
\title{Z-score and p-value for a two-sample proportion test}
\usage{
ai_ztest(NFmin, NPmin, NFmaj, NPmaj)
}
\arguments{
\item{NFmin}{Number of participants failing the selection procedure from the minority group}

\item{NPmin}{Number of participants passing the selection procedure from the minority group}

\item{NFmaj}{number of participants failing the selection procedure from the majority group}

\item{NPmaj}{Number of participants passing the selection procedure from the majority group}
}
\value{
A data frame containing the z-score and associated p-value.
}
\description{
Z-score and p-value for a two-sample proportion test
}
